backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
llama-index-core
llama-index-utils-workflow